XXX maybe more in commit msg: Speed up Encode UTF-8 validation checking
This replaces the current scheme for checking UTF-8 validity by one
in which normal processing doesn't require having to decode the UTF-8
into code points.  The copying of characters individually from the input
to the output is changed to be a single operation for each entire span
of valid input at once.

Thus in the normal case, what ends up happening is a tight loop to
check the validity, and then a memmove of the entire input to the
output, then return.

If an error is found, it copies all the valid input before the error,
then handles the character in error, then positions to the next input
position, and repeats the whole process starting from there.

It uses the functionality available from the Perl 5 core to to look at
just the bytes that comprise the UTF-8 to make the determination,
converting to code points only those that are defective some how in
order to display them in warnings and error messages.

Thus, this does not need to know about the intricacies of UTF-8
malformations, relying on the core to handle this.

This cannot be pushed to CPAN until Devel::PPPort has been updated to
implement all the functions now needed.